edit api doc
